+++ qa/lib/newlib/201-signals.c
@@ -33,7 +33,10 @@ main() {
 	else /* parent */
 	{
 		close(p[1]);
-		read(p[0], &dummy, 1);
+		if(read(p[0], &dummy, 1) > 0){
+			printf("sync read received data");
+			exit(EXIT_FAILURE);
+		}
 		close(p[0]);
 		printf("\nPARENT: sending SIGQUIT\n\n");
 		kill(pid,SIGQUIT);
